requirements.txt
gradio==4.37.2
webvtt-py==0.5.1
yt-dlp @ https://github.com/yt-dlp/yt-dlp/archive/master.tar.gz
